requirements.txt
asgiref==3.8.1
boto3==1.34.102
botocore==1.34.102
dj-database-url==2.1.0
Django==3.2.6
django-admin==2.0.2
django-allauth==0.62.1
django-cors-headers==3.13.0
django-enumchoicefield==2.0.0
django-environ==0.9.0
django-excel-response2==3.0.6
django-extensions==3.2.0
django-filter==22.1
django-heroku==0.3.1
django-multiupload==0.6.1
django-on-heroku==1.1.2
django-ordered-model==3.4.3
django-safedelete==1.3.1
django-secrets-manager==0.1.13
django-six==1.0.5
django-soft-delete==0.9.21
django-softdelete==0.9.4
django-sslify==0.2.7
django-storages==1.11.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.7
excel-base==1.0.4
inflection==0.5.1
isoweek==1.3.3
jmespath==1.0.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.4
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.1
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.11.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
xlwt==1.3.0
